644 fix circular dependency (#645)
* Change location of isRepeatedPassword

* Copy var PASSWORD_REGEX for frontend
