+++ .env.development
@@ -19,7 +19,7 @@ VITE_REACT_APP_API_URL=http://localhost:3000
 
 HASH_ID_SECRET=encode_userid
 
-VITE_PASSWORD_REGEX="^\S*(?=\S{6,})(?=\S*\d)(?=\S*[A-Z])(?=\S*[a-z])(?=\S*[-+!@#$%&*?])\S*$"
+PASSWORD_REGEX="^\S*(?=\S{6,})(?=\S*\d)(?=\S*[A-Z])(?=\S*[a-z])(?=\S*[-+!@#$%&*?])\S*$"
 
 NODEMAILER_HOST=smtp.ethereal.email
 NODEMAILER_PORT=587
@@ -40,3 +40,4 @@ PLAYWRIGHT_WRONG_NAME=
 # FRONT
 VITE_API_URL=http://localhost:10910
 VITE_CHAT_API_URL=http://localhost:10910
+VITE_PASSWORD_REGEX=${PASSWORD_REGEX}
